% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expSolveX.R
\name{expSolveX}
\alias{expSolveX}
\title{Solve for X based on models parameters}
\usage{
expSolveX(y, a, b)
}
\arguments{
\item{y}{value of a y-axis point.}

\item{a}{intercept value of for a \strong{exponential} model.}

\item{b}{growth factor value of for a \strong{exponential} model.}
}
\description{
Given some set of parameters specific to a particular model, solve for x.
}
\examples{
data(wolf)
wolf$Time <- wolf$Year - min(wolf$Year)
e <- expFit(wolf$Time, wolf$Number)
expSolveX(100, e$a, e$b)

}
\seealso{
\code{\link{expFit}}
\code{\link{expFitPred}}
}
\keyword{internal}
