% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.formula.R
\name{is.formula}
\alias{is.formula}
\title{Is the object a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{and object.}
}
\description{
Checks if the object given is a formula notation.
}
\keyword{internal}
