% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linFitPred.R
\name{linFitPred}
\alias{linFitPred}
\title{Linear Fit and plot of data with a predicted value}
\usage{
linFitPred(x, y, xval, xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)), newxlim = NULL, newylim = NULL,
  showpred = TRUE)
}
\arguments{
\item{x}{a designation for a variable on the x-axis. Can be a vector or object.}

\item{y}{a designation for a variable on the y-axis. Can be a vector or object.}

\item{xval}{a value of x that you wish to see a predicted value for.}

\item{xlab}{optional. Text for x-axis title label.}

\item{ylab}{optional. Text for y-axis title label.}

\item{newxlim}{optional. Custom limits for the x axis.}

\item{newylim}{optional. Custom limits for the y axis.}

\item{showpred}{logical. Should the predicted point value be plotted on the graph?

#' @seealso
\code{\link{linFit}}
\code{\link{expFitPred}}
\code{\link{logisticFitPred}}}
}
\description{
Show where a given predicted value will occur, give an input of x.
}
\examples{
data(wolf)
wolf$Time <- wolf$Year - min(wolf$Year)
linFit(wolf$Time, wolf$Number)
linFitPred(wolf$Time, wolf$Number, 6)

}
