% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linSolveX.R
\name{linSolveX}
\alias{linSolveX}
\title{Solve for X based on models parameters}
\usage{
linSolveX(y, a, b)
}
\arguments{
\item{y}{value of a y-axis point (single point).}

\item{a}{intercept value of for a \strong{simple linear} model.}

\item{b}{slope value of for a \strong{simple linear} model.}
}
\description{
Given some set of parameters specific to a particular model, solve for x.
}
\examples{
data(wolf)
wolf$Time <- wolf$Year - min(wolf$Year)
l <- linFit(wolf$Time, wolf$Number)
linSolveX(100, l$Intercept, l$Slope)

}
\seealso{
\code{\link{linFit}}
\code{\link{linFitPred}}
}
\keyword{internal}
