% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticFitPred.R
\name{logisticFitPred}
\alias{logisticFitPred}
\title{Logistic Growth Fit and plot of data with a predicted value}
\usage{
logisticFitPred(
  simpleformula,
  data,
  xval,
  xlab = NULL,
  ylab = NULL,
  newxlim = NULL,
  newylim = NULL,
  showpred = TRUE
)
}
\arguments{
\item{simpleformula}{a designation for the model formula in \code{y ~ x} notation. Must be in formula notation--see details.}

\item{xval}{a value of x that you wish to see a predicted value for.}

\item{xlab}{optional. Text for x-axis title label.}

\item{ylab}{optional. Text for y-axis title label.}

\item{newxlim}{optional. Custom limits for the x axis.}

\item{newylim}{optional. Custom limits for the y axis.}

\item{showpred}{logical. Should the predicted point value be plotted on the graph?}
}
\description{
Show where a given predicted value will occur, give an input of x (logistic growth models).
}
\examples{
data(wolf)
wolf <- wolf \%>\%
  mutate(Time = Year - min(Year))
logisticFit(Number ~ Time, data=wolf)
logisticFitPred(Number ~ Time, data=wolf, xval=6)

}
\seealso{
\code{\link{logisticFit}}
\code{\link{linFitPred}}
\code{\link{expFitPred}}
}
