% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticSolveX.R
\name{logisticSolveX}
\alias{logisticSolveX}
\title{Solve for X based on models parameters}
\usage{
logisticSolveX(y, C, a, b)
}
\arguments{
\item{y}{value of a y-axis point.}

\item{C}{value of for the Carrying Capactiy parameter in a \strong{logistic growth} model.}

\item{a}{value of for the "a" parameter in a \strong{logistic growth} model.}

\item{b}{value of for the "b" parameter in a \strong{logistic growth} model.}
}
\description{
Given some set of parameters specific to a particular model, solve for x.
}
\examples{
data(wolf)
wolf$Time <- wolf$Year - min(wolf$Year)
l <- logisticFit(wolf$Time, wolf$Number)
logisticSolveX(100, l$C, l$a, l$b)

}
\seealso{
\code{\link{logisticFit}}
\code{\link{logisticFitPred}}
}
\keyword{internal}
