% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleFit.R
\name{tripleFit}
\alias{tripleFit}
\title{Fit three models at once}
\usage{
tripleFit(simpleformula, data, xlab = NULL, ylab = NULL)
}
\arguments{
\item{simpleformula}{a designation for the model formula in \code{y ~ x} notation. Must be in formula notation--see details.}

\item{data}{a data frame in which to evaluate formulas.}

\item{xlab}{optional. Text for x-axis title label for produced graph.}

\item{ylab}{optional. Text for y-axis title label for produced graph.}
}
\description{
Fit a linear, exponential, and logistic growth model to two variables and plot the model.
}
\details{
This function is one of a group of functions that take a specific input in the form of \code{y ~ x}, following (in-part) the mosaic package formula notation. However, only simple formulas are allowed with no grouping parameter, nor multiple predictor variables.
}
\examples{
data(wolf)
wolf <- wolf \%>\%
  mutate(Time = Year - min(Year))
tripleFit(Number ~ Time, data=wolf)

}
\seealso{
\code{\link{linFit}}
\code{\link{expFit}}
\code{\link{logisticFit}}
}
