% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expFit.R
\name{expFit}
\alias{expFit}
\title{Exponential Fit and plot of data}
\usage{
expFit(x, y, xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)))
}
\arguments{
\item{x}{a designation for a variable on the x-axis. Can be a vector or object.}

\item{y}{a designation for a variable on the y-axis. Can be a vector or object.}

\item{xlab}{optional. Text for x-axis title label.}

\item{ylab}{optional. Text for y-axis title label.

#' @seealso
\code{\link{expFit}}
\code{\link{logisticFit}}
\code{\link{tripleFit}}}
}
\description{
Fit a exponential model to two variables, return basic output about the model, and plot the model.
}
\examples{
data(wolf)
wolf$Time <- wolf$Year - min(wolf$Year)
expFit(wolf$Time, wolf$Number)
expFit(wolf$Time, wolf$Number, xlab="Years since 1995", ylab="Number of Wolves")
}
\seealso{
\code{\link{linFit}}
\code{\link{logisticFit}}
\code{\link{tripleFit}}
}
