% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleFit.R
\name{tripleFit}
\alias{tripleFit}
\title{Fit three models at once}
\usage{
tripleFit(x, y)
}
\arguments{
\item{x}{a designation for a variable on the x-axis. Must be a vector.}

\item{y}{a designation for a variable on the y-axis. Must be a vector.}
}
\description{
Fit a linear, exponential, and logistic growth model to two variables and plot the model.
}
\examples{
data(wolf)
wolf$Time <- wolf$Year - min(wolf$Year)
tripleFit(wolf$Time, wolf$Number)

}
\seealso{
\code{\link{linFit}}
\code{\link{expFit}}
\code{\link{logisticFit}}
}
